Create user signup form
This commit creates a user signup form and the absolute minimum
necessary framework necessary to get a test for filling it out and
submitting it to work correctly.  Unfortunately, for a simple task, it
takes a full model and two controller actions to work in addition to
one view template that both actions point to.  Not a big deal for a
seasoned rails worker, but a lot for the first-timer.