Test for duplicate usernames
We want to ensure that duplicate usernames are disallowed, but to pass
through any duplicate email addresses in case we want to have the same
address point to multiple accounts.  This requirement could always get
more strict and include only one account/email address, but it is so
easy to create a new email address, why bother?