Don't accept submissions for invalid users.
We add a test in the form submission that ensures the user model
validation is being called correctly.  We simply try submitting an
empty user for now.  This could get more robust, but the user model is
already checking those cases, so full duplication here seems silly.
We just want to make sure the web form doesn't bypass the validation
in some weird manner.