Create tests for validating user.
These tests validate the user at the model and web submission levels.
They ensure that we will not accept an invalid user and require the
user to submit reasonable values for creating a new user.

The interface is not being tested, yet.  We do not ensure that when an
error occurs for these validations being triggered that we pass them
on to the user so that they can know what to do to correct the issue.
That will come later.