Fix up previous test.
After running test with viable candidates, I discovered two errors in
the test that needed to be fixed.  The first was that I was using let
inside a before(:all), which isn't allowed.  The second was that I
wasn't destroying the data created in the before(:all) because they
are created outside a transaction.  Both flaws fixed in this commit.