Test for user index and editing.
These tests are rudimentary tests to see if a user listing can be done
and whether a user can be edited after already in the system.  There
are loads of security flaws that could exist in an implementation that
are not addressed with these tests.  Those will come later after we
can be certain that users can log in to the system.