More theming for bootstrap.
We now have a more consistent colorscheme and a controller that allows
us to run simple visual tests on the screen that can't be automated
since we want to see what the colors look like--not how the web site
performs.

There are probably still colors left default to the bootstrap theme
that should be modified to incorporate the new colorscheme and so more
commits will need to be done.