Initial ability to log in and out
This is a small set of files that will allow us to log in and out as a
user.  It is not complete from testing or functionality, but gets a
lot of the core done without complicating the issues.  More needs to
be added to ensure testing is done properly and that persistent
sessions can be supported safely.