Additional login functionality.
This commit includes functionality to automatically sign in users with
they sign up for the site and then testing to make sure that they are
signed in before they edit their profile page.  The biggest missing
security feature now is the administrator that should be the user
that can delete other accounts.  To appear in a commit soon-ish...