Initial DB design
This commit reflects my first attempt at creating a DB that will work
for the rest of the project along with some very simple tests that
reflect that the DB is implemented properly.  It is expected that
discussion with students over the next several days will result in
modifications to the files since I might be missing a key feature that
they will notice and point out needs to be added to the system.