Don't assume errors come from @user object
This fixes the shared error message so that it can be used by other
models than User by fixing the rather poor assumption that all errors
emanate from the @user object.  Now, the error message partial expects
to be passed a variable named object that contains the error messages
to be displayed.