1:N associations and create contest manager
These tests additionally check for one-to-many associations as well as
contain the initial tests for creating a contest manager.  The contest
manager tests will continue to need development.

As a part of this commit, we are creating code/ directories for
storing the contest managers, but .gitignore updates so those files
don't get included in the repo.  This should ensure that everyone has
the directories created for them, but doesn't try to pass around all
these managers accidentally.