Refactor authentication tests.
Make it easier to say that various tests redirect to root/login
without all the mechanics mucking up the details.  As a result of
refactorization, I found a test that was doing the wrong thing:  it
was trying to update another user instead of deleting that user.  This
patch fixes that bug as well.