Check for delete links properly
This includes a test and implementation that passes those tests that
ensures that the delete links are properly made on the user's show
page.  There was a bug that I wasn't pointing to a player or referee,
so the path was just randomly being set.  Tests idenified this issue
and allowed the fix to be easily added.