Fill out referee tests.
These tests do a good job of completing the referee tests.  They are
quite robust now and try to check for most of the behavior that must
exist for the referee to be correct.  There should be quite a few less
tests for the referee after this commit, since the new tests would
have to deal with more edge cases and not the "normal" path through
the site.