Add title notifier to Guard
This simple line restores the notifier for guard on the title of the
terminal.  This is really nice to keep a long-term record of the last
status of the build since the libnotify pop-up is temporal and goes
away after some amount of time, but the title bar in the terminal is
preserved until the next guard run.