Create new tests for contest model
These tests are the beginning to ensure the contests are created
properly.  There are more functionality tests needed (e.g., ensure two
contests with the same name cannot be added), but the basics are in
place now.  This will provide a nice basis to ensure the contests are
written properly.