More referee and contest validations
These check that we are getting legal values for the contest and
referee models in place, so most of the tests are in the model specs.
The exceptions are when the input from the browser could look legal,
but not be (e.g., picking February 31st).  The model needs to discover
and disallow such options.

We include a gem (validates_timeliness) to help with this, as Rails
decides to make that a legal date by forwarding the appropriate number
of days and we need to munge with Rails internals to get ahold of the
values submitted before they get turned into a date that is valid, but
not what the user chose in the browser.

The implementation included also passes the corresponding tests being
committed.