Fix illegal date test.
The illegal date test was putting the deadline into the future, but
not the start date, so some validations might mark them as being
illegal, but for the wrong reason.  Now we make the start date illegal
since it will be ahead of the deadline and will only catch problems if
the model properly checks the user checkboxes' input correctly.