Final tests for course
This is the final set of tests and code to almost pass them.  There is
one problem with number of players checking in matches that doesn't
work correctly and will need to be modified after I get some sleep.  I
need to commit this, though so that the students have something to
work with.