Match factories not generating valid models
The old factory was not generating valid match models.  The changes
in this commit are required to make them valid, which is odd because
we explicitly turn off validation during creation (see to_create
method override).  The issue here is that while creating the
player_matches for the match, rails needs to know what match_id to
place in the player_match.  To do this, it tries to save the match to
get an id, but the match is not valid yet.  The match needs to have
all the player_matches associated with it before this can be done.  By
turning off validation while we create the player_matches, we can fix
this chicken-and-egg problem.