Refactor FactoryGirl definitions
This code just works and looks a whole lot nicer than the previous
commit.  The only thing I'm unhappy with is the dummy_player value.
It is fragile and could break, but there isn't a way to define a
method in a parent and child at the same callback level, so we do this
to get the best of both worlds.