Allow MatchWrapper to use players with spaces in names
There was a bug in the MatchWrapper that did not quote the player
name on the command line, so any player name with spaces was
improperly sent to the player on startup.  This is a simple fix that
quotes the player name so it appears as a single string to the player
upon startup.