Properly list array of ids in strong parameters
For both the matches and tournament creation, we pass a hash of player
ids that should be added to the created item.  The old code simply did
a query from the DB to get a list of all possible valid ids and
returned an array of integers for the strong parameter filter to use.
This patch modifies the array to be strings since the hash that comes
from the browser stringifies all the values and so the filter was not
passing the ids through properly.  Now this is corrected.