Move from spork to spring.
This commit allows us to use spring that is built in to Rails 4.1
instead of spork as a separate gem.  This involved making some changes
to the Gemfile so that spring could find them all and also installing
the gems into the vendor/cache directory, hence the modification to
the .gitignore file.