Update application for Rails 4.2
These changes allow the code to run using Rails 4.2.0.  The biggest
set of changes is to also upgrade to RSpec 3.1.  This may not have
been necessary, but while chasing down a bug it spec/support/login.rb,
I decided to make the upgrade in case there was an incompatibility
with RSpec 2.14 and Rails 4.2 since the release notes for RSpec 3.1
indicated full support for Rails 4.2.  The bug was more pernicious
than that and may still work on the older version of RSpec, but there
is no point in downgrading when all the specs work with the newer
version.

The major bug fixes included in this commit are:

    1.  The old version of validates_timeliness depended on an older
	implementation of how Rails was assigning values to date and
	time objects.  The change to that is committed in a separate
	repository, but this needs to point to that latest version.
    2.  The "redirects to root" shared example referred to earlier was
	being called in "feature" mode, but depending on "request"
	mode and so the first before action was failing with a
	difficult error to track down.  The solution was to wrap all
	the code in a context that forced everything to be in
	"request" mode and suddenly everything worked properly.

Everything else is explanatory by the diffs.