Update commit to work on CSE VM
We had to update the Gemfile to use a more recent version of eventmachine
because we are using Ruby 2.2 on the CSE VM and it would not work with the
version I had uploaded on my machine.  There was also a change to the
tournament runner script that was somehow not modified in my local commit.

On quick review, everything seems to be working here with the 4.2.0 version
of Rails and so it is ready to merge with the master version to use the
latest and greatest.