Handle uploads without proper names
Since upload=() ties to values during the object build, we can't yet check
whether name is created properly or not.  The two options are:

	1.  Check for name being set before messing with the uploads, or
	2.  Calling upload=() explicitly when we know that everything
	    about the object is valid.

This commit is #1, but #2 is probably more robust.  Unfortunately, it results
in a lot of code churn and seems wrong somehow.  This will probably need to
be revisited during a refactor stage of the code.