Create C++ version of Player and Referee
This is created so that we know what is needed to be able to use the
system from a C++ perspective that requires compilation and such.  It
was created prior to the environments for the Referee being
implemented to ensure that implementation would be satisfactory.  I'm
now just getting to commit it to the repository.